chore: switch to openbao
Hashicorp have switched vault to the BSL license, which is
incompatible with this project's license.

This change switches to openbao, only for the shipped components of
the project. The test suite still relies on hashicorp vault with this
change.

Once openbao is released we should be able to switch fully.

This does not mean we should update any hashicorp components. It
relies on openbao retaining vault compatibility in order to function
with hashicorp vault.